Enabled Route Detail page navigation
